pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pyweste"
version = "0.0.1"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "dearpygui",
    "Pillow",
    "pywin32",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "flake8",
]

[tool.hatch.build.targets.wheel]
packages = ["pyweste"]
